Create Implementing_Google.html
This is an HTML file that will act like Google, using the 'get' method in a form. 

When you search for something on Google, the URL changes from google.com to something with a lot of information after the /. This information is actually a series of parameters that Google uses to create its search results for you. One such parameter is your search query. If you navigate to http://www.google.com/search?q=cats, you’ll notice that the search term "cats" is already filled in for you. q is a key meaning "query" and its value, "cats," is specified after the =.

Let’s implement Google! First, we need an input box for a user’s query:
This code works like the following, where 'q=cats' is the parameter.
Example: http://www.google.com/search?q=cats

Reference Link:
http://cdn.cs50.net/2013/fall/lectures/9/m/notes9m/notes9m.html